Removed class rex_command_package
