+++ redaxo/src/core/lib/console/package/deactivate.php
@@ -1,25 +1,40 @@
 <?php
 
+use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
-use Symfony\Component\Console\Style\SymfonyStyle;
 
 /**
  * @package redaxo\core
  */
-class rex_command_package_deactivate extends rex_command_package
+class rex_command_package_deactivate extends rex_console_command
 {
-    protected function configureCommand()
+    protected function configure()
     {
-        $this->setDescription('Deactivates the selected package');
+        $this
+            ->setDescription('Deactivates the selected package')
+            ->addArgument('package-id', InputArgument::REQUIRED, 'The id of the package (addon or plugin); e.g. "cronjob" or "structure/content"');
     }
 
-    protected function executeCommand(rex_package $package, SymfonyStyle $io, InputInterface $input, OutputInterface $output)
+    protected function execute(InputInterface $input, OutputInterface $output)
     {
+        $io = $this->getStyle($input, $output);
+
+        $packageId = $input->getArgument('package-id');
+        $package = rex_package::get($packageId);
+        if ($package instanceof rex_null_package) {
+            $io->error('Package "'.$packageId.'" doesn\'t exists!');
+            return 1;
+        }
+
         $manager = rex_package_manager::factory($package);
         $success = $manager->deactivate();
-        $message = $manager->getMessage();
-
-        return ['success' => $success, 'message' => $message];
+        $message = strip_tags($manager->getMessage());
+        if ($success) {
+            $io->success($message);
+            return 0;
+        }
+        $io->error($message);
+        return 1;
     }
 }
